can now click to add region
You can only ever have two regions, using the mouse, and it is always a
vertical interface, but you can manipulate the point charges enough that
this is completely general.

I could still add in the ability to rotate the interface, since I have
the general code dealing with it...not sure I should bother though. I'd
rather have the ability to define multiple regions that actually work,
but I think the EFL drawing is currently as slow as is tolerable. (I'm
also not sure how to correctly implement this)