conductors
Adds possibility for a region to be a conductor. This is indicated by eps=0, though I just override the calculation to make it correct in that case.

I also add the possibility of negative permittivity, which should be correct with the same image charge expression except when eps1+ eps2==0, which I just avoid and set things to zero because to treat it properly requires getting into the imaginary parts of the materials' permittivities.
I don't have reliable intuitions about these negative scenarios, and haven't yet found a nice check. I'm just going off of one paper that uses the same image charge expression, so hopefully it is correct.