Adding pdf
Eq 14 and surronding text descrivbe a change in behaviour around a critical distance between two point charges near a negative permittivity material. This situation is not a typical thing to look out for, so I included it because it was neat. I'm not positive but I think I've found the critical behavior change reproduced in EFL viewer.